feat: Enhance test authentication with offline mode, improved error handling, and specific Minecraft/Forge versions, while applying minor formatting fixes.

+++ test/index.js
@@ -5,9 +5,10 @@ const {
   Mojang,
 } = require("@fitzxel/minecraft-java-core");
 const prompt = require("prompt");
-const launcher = new Launch();
-
 const fs = require("fs");
+
+const launcher = new Launch();
+const offline = true;
 const accJson = "./test/account.json";
 let mc;
 
@@ -16,30 +17,43 @@ const client_id = "7ae13d13-d132-462b-8619-bfb13246a563";
 const client_secret = "";
 
 (async () => {
-  const internet = await checkInternet();
-  const ms = internet
-    ? new Microsoft(client_id, undefined, undefined, client_secret)
-    : null;
+  try {
+    if (offline) throw "[#] Offline mode";
 
-  if (!fs.existsSync(accJson)) {
-    mc = await ms.getAuth("raw");
-    fs.writeFileSync(accJson, JSON.stringify(mc, null, 4));
-  } else {
-    mc = JSON.parse(fs.readFileSync(accJson));
-    if (!mc.refresh_token) {
+    const internet = await checkInternet();
+    const ms = new Microsoft(client_id, undefined, undefined, client_secret);
+
+    const fileFound = fs.existsSync(accJson);
+
+    if (!internet && !fileFound) {
+      throw "[#] No internet connection and no account file found.";
+    }
+
+    if (!fileFound) {
+      console.log("[#] No account file found, getting new auth...");
       mc = await ms.getAuth("raw");
       fs.writeFileSync(accJson, JSON.stringify(mc, null, 4));
-    } else if (internet) {
-      mc = await ms.refresh(mc);
-      if (mc.error) mc = await ms.getAuth("raw");
-      fs.writeFileSync(accJson, JSON.stringify(mc, null, 4));
+    } else {
+      mc = JSON.parse(fs.readFileSync(accJson));
+      if (!mc.refresh_token) {
+        console.log("[#] No refresh token found, getting new auth...");
+        mc = await ms.getAuth("raw");
+        fs.writeFileSync(accJson, JSON.stringify(mc, null, 4));
+      } else {
+        console.log("[#] Refreshing auth...");
+        mc = await ms.refresh(mc);
+        if (mc.error) mc = await ms.getAuth("raw");
+        fs.writeFileSync(accJson, JSON.stringify(mc, null, 4));
+      }
     }
-  }
 
-  if (mc) {
+    if (mc.error) throw mc;
     console.log("[#] Authenticated as:", mc.name);
-  } else {
-    console.log("[#] Using offline mode.");
+  } catch (err) {
+    console.log(err);
+    if (err !== "[#] Offline mode")
+      console.log("[#] Auth failed, using offline mode.");
+
     prompt.start();
     const { username } = await prompt.get(["username"]);
     mc = await Mojang.login(username);
@@ -48,12 +62,12 @@ const client_secret = "";
   launcher.config({
     path: "./test/minecraft",
     authenticator: mc,
-    version: "latest_release",
+    version: "1.20.1",
     intelEnabledMac: true,
     instance: "test-instance",
     loader: {
       type: "forge",
-      build: "latest",
+      build: "1.20.1-47.4.10",
       enable: false,
     },
     memory: {
@@ -73,12 +87,14 @@ const client_secret = "";
   launcher
     .on("progress", (progress, size, element) =>
       console.log(
-        `[Game DL]: ${element} - ${((progress / size) * 100).toFixed(2)}%`,
-      ),
+        `[Game DL]: ${element} - ${((progress / size) * 100).toFixed(2)}%`
+      )
     )
     .on("patch", (patch) => process.stdout.write(`[Game Patch]: ${patch}`))
     .on("data", (line) => process.stdout.write(`[Game Data]: ${line}`))
-    .on("error", (error) => console.error(`[Game Error]: ${error.message}`))
+    .on("error", (error) =>
+      console.error(`[Game Error]: ${error.message || JSON.stringify(error)}`)
+    )
     .on("close", (code) => {
       console.log(`[Game]: exited with code ${code}.`);
       process.exit(0);
@@ -92,18 +108,23 @@ const client_secret = "";
     process.exit(0);
   } else {
     const mcProcess = await launcher.start();
-    console.log(`
+    if (mcProcess.pid) {
+      console.log(`
             ---------- Minecraft launched ----------
         [#] Minecraft launched successfully! PID: ${mcProcess.pid}
             ----------------------------------------
         `);
 
-    if (process.argv.includes("--auto-close")) {
-      console.log("[#] Auto-closing Minecraft in 20 seconds...");
-      setTimeout(() => {
-        console.log("[#] Minecraft closed automatically...");
-        mcProcess.kill();
-      }, 20000);
+      if (process.argv.includes("--auto-close")) {
+        console.log("[#] Auto-closing Minecraft in 20 seconds...");
+        setTimeout(() => {
+          console.log("[#] Minecraft closed automatically...");
+          mcProcess.kill();
+        }, 20000);
+      }
+    } else {
+      console.log("[#] Minecraft failed to launch.");
+      process.exit(1);
     }
   }
 })();
